requirements.txt
pandas
numpy
pendulum
python-dotenv
paho-mqtt
boto3